Add ScriptInfo class and make _ucUtils.getScriptData() use it
This new ScriptInfo class is used to share script meta information to
user scripts. Data fields are a copy of the internal representation
(just like before) but we can extend this class to add some helper
methods etc. Right now it only has one .asFile() method to access the
file in question.

This patch also adds a new utils API .parseStringAsScriptInfo() to
that can be used to construct ScriptInfo objects from string.