Make _ucUtils.getScriptData() accept filters and write tests for it
With this patch getScriptData() will now accept two types of filters as
argument:

When called with a string it will return a single ScriptInfo object for a
script which Filename (without path) exactly matches the given string OR
null if no match was found.

When called with a function it will return array of ScriptInfo objects
(even if only one match was found) that match the function.

With no arguments getScriptInfo() behaves just like before.