Parse multi-line description as block-comment
This patch changes multi-line description parsing behavior IF script
header text includes @long-description header.

In such case, the description starts from '/*' sequence on the NEXT line
after @description header and ends at '*/'

In addition, if @long-description header is present, then "normal"
single-line value in @description is ignored, but it can be given a
value to be used as fallback for loaders without multi-line support.