Tests: write few tests for _ucUtils.prefs.{add|remove}Listener
The setup is kinda hairy, because we need to set up listeners before
they would fire, but make it possible to reject and delete them on some
timeout. Thus, we create only one instance of pref listener which lives
for the entire test set (for allowUnsafeWrites which the test toggles
a few times anyway) and create separate tests which give this listener
wrapper a callback to fire when the pref changes. The callback will then
resolve or reject as configured in the test.

This patch also removes separate createTimeoutLong in test runner.
Instead the renamed Test.resolveOnTimeout now accepts millis as argument