Tests: Some improvement for test runner
This patch adds moves a write to sharedGlobal that we use for tests into
separate script - so we can test that the same object is actually shared
among scripts.

In addition, a new "waitable" class is added into test runner which is
used to store the runtime test state of the test set. This allows us to
log info about timed out tests properly.