Make _ucUtils.createWidget throw on failure
Previously we just returned null on failure but it is better to just
throw an Error on unsupported operations.

This changes API behavior a bit such that the method throws when:

id field is nullish
type field is not toolbarbutton or toolbaritem
a widget with same id already exists.

The id and type requirements are essentially the same as before, but
previously we just returned the existing widget. However, the API is
clearly named "create" so it really should be an error if a widget with
same name already exists and the caller should deal with that.