Add support for declarative manifest parsing
This patch adds a new mechanism that can be used to declare a .manifest
file in you script header. The intention here is that the user can
declare a manifest to override some chrome packages for say custom
icons, without needing to write any javascript. Moreover, this allows
declaring a manifest once, while the script itself can be run for every
target document.

The manifest declaration looks simply like this:

// ==UserScript==
// @manifest      my_manifest
// ==/UserScript==

This will cause the loader module to register my_manifest.manifest file
from the scripts directory, but nothing prevents using
"../directory/my_manifest" instead.

In addition, this patch adds two new fields for ScriptData (and thus
ScriptInfo) objects that reflect the manifest name and also if the
script is non-executable. The non-executable field means that the script
file doesn't have any executable javascript content and thus the loader
won't try to inject it to documents. This mechanism simply checks if
the script file-size is bigger than the header size so it's not super
accurate.