esm: separate utils into its own module
This requires a bit of piping because utils needs to know about some
symbols in boot module (such as _ucjs.scripts). To get around that we
run a special method "_linkLoaderModule" when main UserChromeJS is being
initialized which then tells some things to utils.

So, utils is fully initialized and usable before any third-party scripts
are being run and now that it is a module it can be imported to other
sys.mjs scripts if that is desired.