loader: Set regex field to null for background scripts and agent styles
We used to compute and store new RegExp object for them with default
value matching main window. But this was both unnecessary and
misleading.

This leads to slight API changes since now the ScriptInfo object
representing a script or style might have either null or RegExp in
.regex field.

In addition, agent styles now have -1 in loadOrder whereas
previously they had either the default 10 or whatever the user provided.
This is changed because the old configurable value isn't meaningful for
agent styles and wouldn't be respected anyway.