utils: Rewrite utils.registerHotkey and utils.windows helpers
This patch removes old registerHotkey and replaces it with a new Hotkey
class. This class can either be imported from utils or accessed via
utils.hotkeys

instances of Hotkey will return a hotkey descriptor object with two
keys:

command: a description for to-be-constructed <command> element or null
if no command should be created
trigger: a description for to-be-constructed <key> element

Note, simply creating a Hotkey instance will not add it to any window.
You need to either add it manually by using instance method
.attachToWindow(window), or
automatically add it using instance.autoAttach() method which would
create necessary <command> and <key> elements to all current windows as
well as add them to windows that are later opened.

This patch also removes old utils.windows object as well as
utils.windowIsReady method, and replaces them with a
new windowUtils class which provides methods similar to the old one and
some more. The API for utils.windows stays largely the same, notably
utils.windowIsReady() would now be utils.windows.waitWindowLoading()