Loader: replace _ucUtils with UC_API and deprecate @startup
This patch removes the old _ucUtils symbol that was exported to top
level windows with new UC_API symbol which wraps utility methods to a
new nicer interface and can be successfully imported more easily.

This is a breaking change so if you were using _ucUtils then your
scripts need to be updated.

startup directive is also removed since it was very confusing to use,
and UC_API.Windows.onCreated() can be used for the same purpose with
a nicer inteface.

Also version bump