Rework SharedStorage internals
This changes how SharedStorage is implemented, it is now exposed as a
Proxy to a new Storage class instance which provides some new features.

The proxy is used so that *for_time_being* this change should not
require any changes to existing scripts.

The Storage object, exposed via UC_API.SharedStorage, can now add/remove
listeners to storageChange events. This works similarly to how extension
storage API works, with the exception that the methods for writing and
and reading as well as event callbacks are all sychronous unlike their
WE counterparts.