Forward disableClickingJackingProtection argument to appendNotification
As a security feature, Notificationbox buttons don't do anything within
the first Xms after creation of the notification and after the first Xms
of the window becoming focused. Essentially this means that button
doesn't seem to do anything if clicked in a non-focused window.

Using this option disables the security feature for the created
notification.