Loader: Skip initial about:blank window in window created observer
Changes from bug 543435 cause the first DOMWindowCreated event to
be dispatched for about:blank document. If we detect that, then we now
add another DOMContent listener which then detects the following
actual window DOMContent event. The reason for this is that for now
we want to use { once: true } option when adding the event listener.
This fixes #87